refactor: re-organize portal-bridge
