+++ portal-bridge/src/api/execution.rs
@@ -1,12 +1,13 @@
-use crate::full_header::{FullHeader, FullHeaderBatch};
-use crate::pandaops::PandaOpsMiddleware;
 use anyhow::{anyhow, bail};
 use ethereum_types::H256;
+use serde_json::{json, Value};
+
+use crate::pandaops::PandaOpsMiddleware;
+use crate::types::full_header::{FullHeader, FullHeaderBatch};
 use ethportal_api::types::jsonrpc::params::Params;
 use ethportal_api::types::jsonrpc::request::JsonRequest;
 use ethportal_api::utils::bytes::hex_encode;
 use ethportal_api::{Header, Receipts};
-use serde_json::{json, Value};
 
 /// Implements endpoints from the Execution API to access data from the execution layer.
 #[derive(Clone, Debug, PartialEq, Eq)]
