+++ portal-bridge/src/bridge/mod.rs
@@ -1,0 +1,2 @@
+pub mod beacon;
+pub mod history;
