+++ portal-bridge/src/cli.rs
@@ -7,8 +7,8 @@ use tokio::process::Child;
 use url::Url;
 
 use crate::client_handles::{fluffy_handle, trin_handle};
-use crate::mode::BridgeMode;
-use crate::types::NetworkKind;
+use crate::types::mode::BridgeMode;
+use crate::types::network::NetworkKind;
 use ethportal_api::types::cli::check_private_key_length;
 
 // max value of 16 b/c...
@@ -128,7 +128,7 @@ impl ClientType {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::mode::ModeType;
+    use crate::types::mode::ModeType;
 
     #[test]
     fn test_default_bridge_config() {
