+++ portal-bridge/src/main.rs
@@ -1,16 +1,16 @@
+use std::sync::Arc;
+
 use clap::Parser;
+use tokio::time::{sleep, Duration};
+
 use ethportal_api::jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
 use ethportal_api::types::cli::{DEFAULT_DISCOVERY_PORT, DEFAULT_WEB3_HTTP_PORT};
-use portal_bridge::beacon_bridge::BeaconBridge;
-use portal_bridge::bridge::Bridge;
+use portal_bridge::api::{consensus::ConsensusApi, execution::ExecutionApi};
+use portal_bridge::bridge::{beacon::BeaconBridge, history::HistoryBridge};
 use portal_bridge::cli::BridgeConfig;
-use portal_bridge::consensus_api::ConsensusApi;
-use portal_bridge::execution_api::ExecutionApi;
 use portal_bridge::pandaops::PandaOpsMiddleware;
-use portal_bridge::types::NetworkKind;
+use portal_bridge::types::network::NetworkKind;
 use portal_bridge::utils::generate_spaced_private_keys;
-use std::sync::Arc;
-use tokio::time::{sleep, Duration};
 use trin_utils::log::init_tracing_logger;
 use trin_validation::accumulator::MasterAccumulator;
 use trin_validation::oracle::HeaderOracle;
@@ -78,7 +78,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             let pandaops_middleware = PandaOpsMiddleware::default();
             let execution_api = ExecutionApi::new(pandaops_middleware);
 
-            let bridge = Bridge::new(
+            let bridge = HistoryBridge::new(
                 bridge_config.mode,
                 execution_api,
                 portal_clients.expect("Failed to create history JSON-RPC clients"),
