+++ portal-bridge/src/types/network.rs
